consolidate mattermost and zendesk clients into one directory instead of separate individual directories
