+++ src/clients/index.ts
@@ -1,0 +1,2 @@
+export * as mattermost from './mattermost';
+export * as zendesk from './zendesk';
