+++ src/clients/mattermost.ts
@@ -1,8 +1,8 @@
 import Client from 'mattermost-redux/client/client4.js';
 
-export function newClient(token: string, url: string): Client {
+export const newClient = (token: string, url: string): Client => {
     const client = new Client();
     client.setUrl(url);
     client.setToken(token);
     return client;
-}
+};
